requirements.txt
streamlit
requests
google-generativeai
pandas
toml